fix: add explicit dependency on rigetti-pyo3's "time" feature (#476)
In rigetti/rigetti-pyo3#45 (and rigetti/rigetti-pyo3#46), I moved all
references to the Python time types under the "time" feature.  This
was a bugfix, but I didn't realize it was also a breaking change – any
library that relied on the time types implicitly being available would
break, just as this one did!  This adds the missing feature
dependency, and I've also upgraded rigetti/rigetti-pyo3#46 to be a
breaking change.

This commit also updates Cargo.lock to force the update to
rigetti-pyo3, meaning that builds will fail without the substantive
change.

Closes #475.